st-texture-cache: Load a sized texture taking care of the scaling
The icon size should be adapted so that it will match the final icon
size that we actually want for current floating scaling, without
wasting using a bigger size or having to do unoptimal downscaling
(for quality).

https://bugzilla.gnome.org/show_bug.cgi?id=765011

+++ src/st/st-texture-cache.h
@@ -71,7 +71,7 @@ ClutterActor *st_texture_cache_load_gicon (StTextureCache *cache,
                                            GIcon          *icon,
                                            gint            size,
                                            gint            paint_scale,
-                                           gint            resource_scale);
+                                           gfloat          resource_scale);
 
 ClutterActor *st_texture_cache_load_file_async (StTextureCache    *cache,
                                                 GFile             *file,
